requirements.txt
#
# Requirements without Version Specifiers
setuptools
pytest
Pillow
flake8
black
mypy
